- fixes to the generic graph a*, improved the test grid graph, added a clipped point-to-point path utility

+++ open-tracking-tools-api/src/main/java/org/opentrackingtools/statistics/filters/vehicles/VehicleTrackingFilter.java
@@ -12,7 +12,7 @@ public interface VehicleTrackingFilter<ObservationType, ParameterType>
   FilterInformation getFilterInformation(
     ObservationType observation);
 
-  double getLastProcessedTime();
+  Long getLastProcessedTime();
 
   @Override
   Random getRandom();
